Move installation dialog to a separate class
