+++ package/system-role-common-criteria.spec
@@ -73,6 +73,8 @@ mkdir -p $RPM_BUILD_ROOT/%{_prefix}/share/doc/packages/%{name}
 install -m 644 LICENSE $RPM_BUILD_ROOT/%{_prefix}/share/doc/packages/%{name}
 mkdir -p %{buildroot}/%{_datadir}/YaST2/lib/y2system_role_handlers
 install -m 644 src/lib/y2system_role_handlers/cc_role_finish.rb %{buildroot}/%{_datadir}/YaST2/lib/y2system_role_handlers
+mkdir -p %{buildroot}/%{_datadir}/YaST2/lib/y2common_criteria/dialogs
+install -m 644 src/lib/y2common_criteria/dialogs/installation.rb %{buildroot}/%{_datadir}/YaST2/lib/y2common_criteria/dialogs/installation.rb
 mkdir -p %{buildroot}/%{_datadir}/YaST2/clients
 install -m 644 src/clients/inst_cc_mode.rb %{buildroot}/%{_datadir}/YaST2/clients/inst_cc_mode.rb
